requirements.txt
asgiref==3.6.0
Django==5.1.1
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
protobuf==4.21.12
PyJWT==2.6.0
pytz==2022.7
six==1.16.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
